Call GetLootSlotInfo with the correct return values
The new API introduces the parameter currencyID and, of course, Blizzard
choose to put it the middle of the old parameters just to mess things
up.
This also explains why the locked wasn't working.
This commit brings the locked case back and also adds a currency case
with the new parameter.